Struct layout_thread_2013::LayoutThread

pub struct LayoutThread {

    pub(crate) id: PipelineId,
    pub(crate) url: ServoUrl,
    pub(crate) stylist: Stylist,
    pub(crate) is_iframe: bool,
    pub(crate) font_cache_sender: IpcSender<()>,
    pub(crate) constellation_chan: IpcSender<LayoutMsg>,
    pub(crate) script_chan: IpcSender<ConstellationControlMsg>,
    pub(crate) time_profiler_chan: ProfilerChan,
    pub(crate) image_cache: Arc<dyn ImageCache>,
    pub(crate) font_cache_thread: FontCacheThread,
    pub(crate) first_reflow: Cell<bool>,
    pub(crate) parallel_flag: bool,
    pub(crate) generation: Cell<u32>,
    pub(crate) outstanding_web_fonts: Arc<AtomicUsize>,
    pub(crate) root_flow: RefCell<Option<FlowRef>>,
    pub(crate) epoch: Cell<Epoch>,
    pub(crate) viewport_size: Size2D<Au>,
    pub(crate) rw_data: Arc<Mutex<LayoutThreadData>>,
    pub(crate) webrender_image_cache: Arc<RwLock<FnvHashMap<(ServoUrl, UsePlaceholder), WebRenderImageInfo>>>,
    pub(crate) registered_painters: RegisteredPaintersImpl,
    pub(crate) webrender_api: WebrenderIpcSender,
    pub(crate) paint_time_metrics: PaintTimeMetrics,
    pub(crate) layout_query_waiting_time: Histogram,
    pub(crate) last_iframe_sizes: RefCell<FnvHashMap<BrowsingContextId, Size2D<f32, CSSPixel>>>,
    pub(crate) debug: DebugOptions,
    pub(crate) nonincremental_layout: bool,

}

Information needed by layout.

Fields
id: PipelineId

The ID of the pipeline that we belong to.

url: ServoUrl

The URL of the pipeline that we belong to.

stylist: Stylist

Performs CSS selector matching and style resolution.

is_iframe: bool

Is the current reflow of an iframe, as opposed to a root window?

font_cache_sender: IpcSender<()>

The channel on which the font cache can send messages to us.

constellation_chan: IpcSender<LayoutMsg>

The channel on which messages can be sent to the constellation.

script_chan: IpcSender<ConstellationControlMsg>

The channel on which messages can be sent to the script thread.

time_profiler_chan: ProfilerChan

The channel on which messages can be sent to the time profiler.

image_cache: Arc<dyn ImageCache>

Reference to the script thread image cache.

font_cache_thread: FontCacheThread

Public interface to the font cache thread.

first_reflow: Cell<bool>

Is this the first reflow in this LayoutThread?

parallel_flag: bool

Flag to indicate whether to use parallel operations

generation: Cell<u32>

Starts at zero, and increased by one every time a layout completes. This can be used to easily check for invalid stale data.

outstanding_web_fonts: Arc<AtomicUsize>

The number of Web fonts that have been requested but not yet loaded.

root_flow: RefCell<Option<FlowRef>>

The root of the flow tree.

epoch: Cell<Epoch>

A counter for epoch messages

viewport_size: Size2D<Au>

The size of the viewport. This may be different from the size of the screen due to viewport constraints.

rw_data: Arc<Mutex<LayoutThreadData>>

A mutex to allow for fast, read-only RPC of layout’s internal data structures, while still letting the LayoutThread modify them.

All the other elements of this struct are read-only.

webrender_image_cache: Arc<RwLock<FnvHashMap<(ServoUrl, UsePlaceholder), WebRenderImageInfo>>>registered_painters: RegisteredPaintersImpl

The executors for paint worklets.

webrender_api: WebrenderIpcSender

Webrender interface.

paint_time_metrics: PaintTimeMetrics

Paint time metrics.

layout_query_waiting_time: Histogram

The time a layout query has waited before serviced by layout.

last_iframe_sizes: RefCell<FnvHashMap<BrowsingContextId, Size2D<f32, CSSPixel>>>

The sizes of all iframes encountered during the last layout operation.

debug: DebugOptions

Debug options, copied from configuration to this LayoutThread in order to avoid having to constantly access the thread-safe global options.

nonincremental_layout: bool

True to turn off incremental layout.

Implementations

impl LayoutThread

pub(crate) fn new( id: PipelineId, url: ServoUrl, is_iframe: bool, constellation_chan: IpcSender<ConstellationMsg>, script_chan: IpcSender<ConstellationControlMsg>, image_cache: Arc<dyn ImageCache>, font_cache_thread: FontCacheThread, time_profiler_chan: ProfilerChan, webrender_api: WebrenderIpcSender, paint_time_metrics: PaintTimeMetrics, window_size: WindowSizeData ) -> LayoutThread

pub(crate) fn build_layout_context<'a>( &'a self, guards: StylesheetGuards<'a>, snapshot_map: &'a SnapshotMap, origin: ImmutableOrigin, animation_timeline_value: f64, animations: &DocumentAnimationSet, stylesheets_changed: bool ) -> LayoutContext<'a>

pub(crate) fn handle_request<'a, 'b>(&mut self, request: Request)

Receives and dispatches messages from the script and constellation threads

pub(crate) fn handle_request_helper<'a, 'b>( &mut self, request: Msg, possibly_locked_rw_data: &mut RwData<'a, 'b> )

Receives and dispatches messages from other threads.

pub(crate) fn collect_reports<'a, 'b>( &self, reports_chan: ReportsChan, possibly_locked_rw_data: &mut RwData<'a, 'b> )

pub(crate) fn exit_now(&mut self)

Shuts down layout now.

pub(crate) fn load_all_web_fonts_from_stylesheet_with_guard( &self, stylesheet: &Stylesheet, guard: &SharedRwLockReadGuard<'_> )

pub(crate) fn handle_web_font_loaded(&self)

pub(crate) fn handle_set_quirks_mode<'a, 'b>(&mut self, quirks_mode: QuirksMode)

Sets quirks mode for the document, causing the quirks mode stylesheet to be used.

pub(crate) fn try_get_layout_root<'dom>( &self, node: impl LayoutNode<'dom> ) -> Option<FlowRef>

pub(crate) fn solve_constraints( layout_root: &mut dyn Flow, layout_context: &LayoutContext<'_> )

Performs layout constraint solving.

This corresponds to Reflow() in Gecko and layout() in WebKit/Blink and should be benchmarked against those two. It is marked #[inline(never)] to aid profiling.

pub(crate) fn solve_constraints_parallel( traversal: &ThreadPool, layout_root: &mut dyn Flow, profiler_metadata: Option<TimerMetadata>, time_profiler_chan: ProfilerChan, layout_context: &LayoutContext<'_> )

Performs layout constraint solving in parallel.

This corresponds to Reflow() in Gecko and layout() in WebKit/Blink and should be benchmarked against those two. It is marked #[inline(never)] to aid profiling.

pub(crate) fn update_iframe_sizes( &self, new_iframe_sizes: FnvHashMap<BrowsingContextId, Size2D<f32, CSSPixel>> )

Update the recorded iframe sizes of the contents of layout and when these sizes changes, send a message to the constellation informing it of the new sizes.

pub(crate) fn compute_abs_pos_and_build_display_list( &self, data: &Reflow, reflow_goal: &ReflowGoal, document: Option<&ServoLayoutDocument<'_, LayoutData>>, layout_root: &mut dyn Flow, layout_context: &mut LayoutContext<'_>, rw_data: &mut LayoutThreadData )

Computes the stacking-relative positions of all flows and, if the painting is dirty and the reflow type need it, builds the display list.

pub(crate) fn handle_reflow<'a, 'b>( &mut self, data: &mut ScriptReflowResult, possibly_locked_rw_data: &mut RwData<'a, 'b> )

The high-level routine that performs layout.

pub(crate) fn respond_to_query_if_necessary( &self, reflow_goal: &ReflowGoal, rw_data: &mut LayoutThreadData, context: &mut LayoutContext<'_>, reflow_result: &mut ReflowComplete, shared_lock: &SharedRwLock )

pub(crate) fn update_scroll_node_state( &self, state: &ScrollState, rw_data: &mut LayoutThreadData )

pub(crate) fn set_scroll_states<'a, 'b>( &mut self, new_scroll_states: Vec<ScrollState>, possibly_locked_rw_data: &mut RwData<'a, 'b> )

pub(crate) fn cancel_animations_for_nodes_not_in_flow_tree( animations: &mut FxHashMap<AnimationSetKey, ElementAnimationSet>, root_flow: &mut dyn Flow )

Cancel animations for any nodes which have been removed from flow tree. TODO(mrobinson): We should look into a way of doing this during flow tree construction. This also doesn’t yet handles nodes that have been reparented.

pub(crate) fn perform_post_style_recalc_layout_passes( &self, root_flow: &mut FlowRef, data: &Reflow, reflow_goal: &ReflowGoal, document: Option<&ServoLayoutDocument<'_, LayoutData>>, rw_data: &mut LayoutThreadData, context: &mut LayoutContext<'_>, thread_pool: Option<&ThreadPool> )

pub(crate) fn perform_post_main_layout_passes( &self, data: &Reflow, root_flow: &mut FlowRef, reflow_goal: &ReflowGoal, document: Option<&ServoLayoutDocument<'_, LayoutData>>, rw_data: &mut LayoutThreadData, layout_context: &mut LayoutContext<'_> )

pub(crate) fn reflow_all_nodes(flow: &mut dyn Flow)

pub(crate) fn profiler_metadata(&self) -> Option<TimerMetadata>

Returns profiling information which is passed to the time profiler.

pub(crate) fn handle_viewport_change( &mut self, window_size_data: WindowSizeData, guards: &StylesheetGuards<'_> ) -> bool

Update layout given a new viewport. Returns true if the viewport changed or false if it didn’t.